Fixed NPE on toLowerCase calls for Entity Names & Dropped SilkSpawners 5.x support
