+++ pom.xml
@@ -63,7 +63,7 @@
         <dependency>
             <groupId>de.dustplanet</groupId>
             <artifactId>SilkSpawners</artifactId>
-            <version>5.0.2</version>
+            <version>8.0.0</version>
         </dependency>
     </dependencies>
 </project>
