+++ src/main/java/net/brcdev/shopgui/bridge/silkspawners/ShopGuiBridgeSilkSpawnersPlugin.java
@@ -1,11 +1,8 @@
 package net.brcdev.shopgui.bridge.silkspawners;
 
 import net.brcdev.shopgui.ShopGuiPlusApi;
-import net.brcdev.shopgui.bridge.silkspawners.spawner.SilkSpawnersLegacyProvider;
-import net.brcdev.shopgui.bridge.silkspawners.spawner.SilkSpawnersCurrentProvider;
 import net.brcdev.shopgui.bridge.silkspawners.spawner.SilkSpawnersProvider;
 import net.brcdev.shopgui.exception.api.ExternalSpawnerProviderNameConflictException;
-import org.bukkit.Bukkit;
 import org.bukkit.plugin.java.JavaPlugin;
 
 public class ShopGuiBridgeSilkSpawnersPlugin extends JavaPlugin {
@@ -18,12 +15,8 @@ public class ShopGuiBridgeSilkSpawnersPlugin extends JavaPlugin {
   }
 
   private void hookIntoSilkSpawners() {
-    if (usingLegacySilkSpawners()) {
-      this.spawnerProvider = new SilkSpawnersLegacyProvider();
-    } else {
-      this.spawnerProvider = new SilkSpawnersCurrentProvider();
-    }
-    this.spawnerProvider.hookIntoSilkSpawners(Bukkit.getPluginManager().getPlugin("SilkSpawners"));
+    this.spawnerProvider = new SilkSpawnersProvider();
+    this.spawnerProvider.hookIntoSilkSpawners();
   }
 
   private void hookIntoShopGui() {
@@ -32,17 +25,6 @@ public class ShopGuiBridgeSilkSpawnersPlugin extends JavaPlugin {
     } catch (ExternalSpawnerProviderNameConflictException e) {
       getLogger().warning("Failed to hook into ShopGUI+: " + e.getMessage());
     }
-  }
-
-  private boolean usingLegacySilkSpawners() {
-    int versionMajorNumber;
-    String version = Bukkit.getPluginManager().getPlugin("SilkSpawners").getDescription().getVersion();
-    try {
-      versionMajorNumber = Integer.parseInt(String.valueOf(version.charAt(0)));
-    } catch (NumberFormatException | ArrayIndexOutOfBoundsException ex) {
-      return true;
-    }
-    return (versionMajorNumber < 6);
   }
 }
